refactor(fsengage): Refactor fsengage to only include abstraction layer
BREAKING CHANGE: Vendor-specific implementations are now separate modules
