+++ packages/fsengage/package.json
@@ -13,14 +13,10 @@
   "dependencies": {
     "@brandingbrand/fsfoundation": "^3.0.0",
     "@brandingbrand/fsnetwork": "^3.0.0",
-    "@brandingbrand/react-native-adobe-marketing-cloud": "^0.1.1",
-    "@brandingbrand/react-native-google-analytics": "^2.1.0",
-    "@brandingbrand/react-native-leanplum": "^1.0.1",
     "decimal.js": "^10.0.0",
     "react": "16.6.3",
     "react-native": "0.57.8",
-    "react-native-permissions": "^1.1.1",
-    "url-parse": "^1.4.3"
+    "react-native-permissions": "^1.1.1"
   },
   "devDependencies": {
     "sinon": "^7.3.0"
