+++ packages/fsengage/src/modules/analytics/index.ts
@@ -1,2 +1,8 @@
-export * from './providers';
+import * as AnalyticsProvider from './AnalyticsProvider';
+import AnalyticsProviderConfiguration from './AnalyticsProviderConfiguration';
+
 export { default as Analytics } from './Analytics';
+export {
+  AnalyticsProvider,
+  AnalyticsProviderConfiguration
+};
