--- packages/fsengage/src/modules/analytics/providers/index.ts
@@ -1,6 +1,0 @@
-export { default as AnalyticsProvider } from './AnalyticsProvider';
-
-export { default as AdobeMarketingCloudProvider } from './adobe/AdobeMarketingCloudProvider';
-export { default as GoogleAnalyticsProvider } from './google/GoogleAnalyticsProvider';
-export { default as LeanplumProvider } from './leanplum/LeanplumProvider';
-export * from './adobe/AdobeAnalyticsProvider';
