--- packages/fsengage/src/modules/analytics/providers/types/index.ts
@@ -1,1 +1,0 @@
-export { default as AnalyticsProviderConfiguration } from './AnalyticsProviderConfiguration';
