--- packages/fsengage/src/typings.d.ts
@@ -1,1 +1,0 @@
-declare module '@brandingbrand/react-native-leanplum';
